+++ packages/ctl/index.ts
@@ -1,0 +1,6 @@
+import { bootstrap } from "./ctl";
+export { default as BaseCtlPlugin } from "./src/BaseCtlPlugin";
+
+if (module === require.main) {
+	bootstrap();
+}
