+++ packages/ctl/package.json
@@ -4,7 +4,7 @@
 	"author": "Hornwitser",
 	"license": "MIT",
 	"version": "2.0.0-alpha.13",
-	"main": "dist/ctl.js",
+	"main": "dist/index.js",
 	"scripts": {
 		"prepare": "tsc --build",
 		"test": "echo \"Error: no test specified\" && exit 1"
