+++ packages/host/host.ts
@@ -245,7 +245,7 @@ async function startHost() {
 	logger.info("Started host");
 }
 
-if (module === require.main) {
+export function bootstrap() {
 	// eslint-disable-next-line no-console
 	console.warn(`
 +==========================================================+
@@ -286,4 +286,8 @@ ${err.stack}`
 
 		process.exitCode = 1;
 	});
+}
+
+if (module === require.main) {
+	bootstrap();
 }
