+++ packages/host/index.ts
@@ -1,0 +1,6 @@
+import { bootstrap } from "./host";
+export { default as BaseInstancePlugin } from "./src/BaseInstancePlugin";
+
+if (module === require.main) {
+	bootstrap();
+}
