+++ packages/host/tsconfig.json
@@ -3,7 +3,7 @@
 	"references": [
 		{ "path": "../lib" },
 	],
-	"include": ["host.ts", "package.json", "src/**/*.ts", "src/**/*.js"],
+	"include": ["index.ts", "host.ts", "package.json", "src/**/*.ts", "src/**/*.js"],
 	"compilerOptions": {
 		"outDir": "dist",
 	},
